VIX-3155 Separate Classes in ColorBreakdownModule.cs into separate cs files
